requirements.txt
PYJWT
flask
flask_sqlalchemy
Flask_Reuploaded
